1.3.4 RELEASE, This is probably one of my most important releases, this is the last release using TinyProtocol!
Changelog:
* Clientbound Abilities wrapper added and is sendable to clients

* Clientbound Health update wrapper added and is also sendable to clients

Info:
TinyProtocol helped us listen to some login and handshake packets!
But it also came with a lot of bugs and problems to deal with, even today there are still some bugs and many people using this API have complained about it.

I will now stop using TinyProtocol.

WHAT WILL THIS MEAN?

There will be lots of pro's and con's.

Pros:
* No more issues with reloading
* Smaller file size

Cons:
* Version lookup will now only be possible with dependencies like ViaVersion.
* Soft depend on via version as mentioned above. Soft depend meaning that if you want to lookup versions, ViaVersion api is required to be provided.

* All login wrappers and the login event will be removed!

It is very important that PacketEvents is stable and people shouldn't be constantly running into issues!
I am releasing this version with TinyProtocol for the users that will miss and still want TinyProtocol.