To summarize this push, we rewrote the outgoing chat wrapper.
PacketWrapper appropriately named methods that have to do with readLastSeenMessages and added the true readLastSeenMessages;

WrapperPlayServerChatMessage had many changes in 1.19.1 so we decided to go with a different approach of implementing it and decided to go with a recode.
We have chat processors (readers & writers) which implement readers and writers for various versions. We have a chatmessage class which contains fields available on every single version and has subclasses which contain fields for their own version. Cast the chat message to get fields for specific versions. (Kinda like how BaseChunks are implemented)