User Storage System
Added a system that allows developers to associate a certain value with the PacketEvent User class, making it easier and faster to identify the user on any PacketEvent at all times.
MethodHandle was used instead of VarHandle, as I found it to be slightly faster, with, and without warmup calls.
Here's the benchmark with 1000 warmup calls and 10000 evaluation calls: https://paste.md-5.net/